# companion/openapi.yaml
openapi: 3.1.0
info:
  title: Companion API
  version: 0.1.0
  description: |
    API del empleado IA central del ecosistema.

    Companion acepta API key o Bearer JWT interno. Cuando hay principal
    autenticado, `org_id` se interpreta como customer org/work context: la
    empresa cliente donde trabaja el empleado IA. Headers como `X-Org-ID` y
    `X-User-ID` quedan como compatibilidad de dev/test, no como fuente canonica
    productiva.
servers:
  - url: http://localhost:18085
security:
  - apiKeyAuth: []
  - bearerAuth: []
paths:
...
  /v1/memories:
    get:
      summary: Lista contenedores Memory
      description: Memory es el contenedor fuerte que puede referenciar un Virtual Employee mediante `memory_id`.
      parameters:
        - in: query
          name: lifecycle
          schema: { type: string, enum: [active, archived, all], default: active }
      responses:
        "200":
          description: Memory container list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Memory" }
    post:
      summary: Crea un contenedor Memory
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemoryInput" }
      responses:
        "201":
          description: Memory container created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Memory" }
  /v1/memories/{memory_id}:
    get:
      summary: Obtiene un contenedor Memory
      parameters:
        - in: path
          name: memory_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Memory container
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Memory" }
    patch:
      summary: Actualiza un contenedor Memory
      parameters:
        - in: path
          name: memory_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemoryInput" }
      responses:
        "200":
          description: Memory container updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Memory" }
  /v1/memories/{memory_id}/status:
    post:
      summary: Cambia status de un contenedor Memory
      parameters:
        - in: path
          name: memory_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, disabled, archived] }
      responses:
        "200":
          description: Memory status changed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Memory" }
  /v1/memories/{memory_id}/entries:
    get:
      summary: Lista entries de un contenedor Memory
      parameters:
        - in: path
          name: memory_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Memory entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/MemoryContainerEntry" }
    post:
      summary: Crea un entry dentro de un contenedor Memory
      parameters:
        - in: path
          name: memory_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemoryContainerEntryInput" }
      responses:
        "201":
          description: Memory entry created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MemoryContainerEntry" }
...
  /v1/virtual-employees:
    get:
      summary: Lista Virtual Employees de la customer org
      description: API publica recomendada para trabajadores digitales persistentes. `tenant_id` es UUID publico y Companion mantiene `org_id + product_surface` solo como contexto interno de transicion.
      parameters:
        - in: query
          name: lifecycle
          schema: { type: string, enum: [active, archived, trashed, all], default: active }
      responses:
        "200":
          description: Virtual Employee list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/VirtualEmployee" }
    post:
      summary: Crea un Virtual Employee
      description: Crea un trabajador digital persistente con ID automatico. No acepta `agent_id`, `metadata`, `job_title`, `mission`, `responsibilities`, `allowed_tools`, `memory_enabled` ni `memory_scope_id`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VirtualEmployeeInput" }
      responses:
        "201":
          description: Virtual Employee created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VirtualEmployee" }
...
  /v1/virtual-employees/{employee_id}:
    get:
      summary: Obtiene un Virtual Employee
      description: API publica recomendada. `employee_id` es UUID publico del Virtual Employee.
      parameters:
        - in: path
          name: employee_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Virtual Employee
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VirtualEmployee" }
    patch:
      summary: Actualiza un Virtual Employee
      description: Actualiza solo el core del Employee. No cambia status; para lifecycle usar `/status`.
      parameters:
        - in: path
          name: employee_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VirtualEmployeeInput" }
      responses:
        "200":
          description: Virtual Employee updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VirtualEmployee" }
  /v1/virtual-employees/{employee_id}/status:
    post:
      summary: Cambia el status de un Virtual Employee
      parameters:
        - in: path
          name: employee_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: "#/components/schemas/EmployeeStatus" }
      responses:
        "200":
          description: Virtual Employee status changed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VirtualEmployee" }
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Memory:
      type: object
      required: [memory_id, tenant_id, policy, status]
      properties:
        memory_id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        owner_employee_id: { type: string, format: uuid, nullable: true }
        policy: { $ref: "#/components/schemas/MemoryPolicy" }
        status: { type: string, enum: [active, disabled, archived] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        archived_at: { type: string, format: date-time, nullable: true }
        version: { type: integer }
    MemoryInput:
      type: object
      additionalProperties: false
      properties:
        owner_employee_id: { type: string, format: uuid, nullable: true }
        policy: { $ref: "#/components/schemas/MemoryPolicy" }
        status: { type: string, enum: [active, disabled] }
    MemoryPolicy:
      type: object
      required: [enabled_by_default, retention_days, allow_user_memory, allow_task_memory, allow_tenant_memory]
      properties:
        enabled_by_default: { type: boolean }
        retention_days: { type: integer, minimum: 0 }
        allow_user_memory: { type: boolean }
        allow_task_memory: { type: boolean }
        allow_tenant_memory: { type: boolean }
    MemoryContainerEntry:
      type: object
      required: [memory_entry_id, memory_id, kind, content_text, confidence, status]
      properties:
        memory_entry_id: { type: string, format: uuid }
        memory_id: { type: string, format: uuid }
        kind: { type: string, enum: [fact, preference, summary, procedure, note] }
        content_text: { type: string }
        confidence: { type: number, minimum: 0, maximum: 1 }
        status: { type: string, enum: [active, archived] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    MemoryContainerEntryInput:
      type: object
      required: [kind, content_text]
      additionalProperties: false
      properties:
        kind: { type: string, enum: [fact, preference, summary, procedure, note] }
        content_text: { type: string }
        confidence: { type: number, minimum: 0, maximum: 1 }
...
    VirtualEmployee:
      type: object
      description: Trabajador digital persistente con identidad propia. No es un Agent ni un alias de modulo tecnico de agents.
      required: [employee_id, tenant_id, name, supervisor_user_id, status, job_role_id, profile_id, autonomy, capability_ids]
      properties:
        employee_id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        name: { type: string }
        supervisor_user_id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/EmployeeStatus" }
        job_role_id: { type: string, format: uuid }
        profile_id: { type: string, format: uuid }
        autonomy: { $ref: "#/components/schemas/AutonomyLevel" }
        capability_ids:
          type: array
          items: { type: string, format: uuid }
        memory_id: { type: string, format: uuid, nullable: true }
    VirtualEmployeeInput:
      type: object
      required: [name, supervisor_user_id, job_role_id, profile_id, autonomy]
      additionalProperties: false
      properties:
        name: { type: string }
        supervisor_user_id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/EmployeeStatus" }
        job_role_id: { type: string, format: uuid }
        profile_id: { type: string, format: uuid }
        autonomy: { $ref: "#/components/schemas/AutonomyLevel" }
        capability_ids:
          type: array
          items: { type: string, format: uuid }
        memory_id: { type: string, format: uuid, nullable: true }
    EmployeeStatus:
      type: string
      enum: [draft, active, disabled, suspended, archived, trashed, error]
    AutonomyLevel:
      type: string
      enum: [A0, A1, A2, A3, A4, A5]
...